requirements.txt
requests>=2.28.0
gitpython>=3.1.0
pystray>=0.19.0
Pillow>=9.0.0
schedule>=1.2.0
customtkinter>=5.2.0
